+++ docs/source/changes.rst
@@ -11,7 +11,7 @@ chronological order. All previous releases should still be available on pip.
 .. _v0.12.0:
 
 ----------------------
-0.12.0rc1 - 2023-10-23
+0.12.0rc2 - 2023-11-23
 ----------------------
 
 
@@ -61,6 +61,12 @@ Here is a stripped-down example.
 
 
 For more details and examples, see :pull:`553`.
+
+Improvements
+------------
+- :func:`~hydra_zen.BuildsFn` was introduced to permit customizable auto-config and type-refinement support in config-creation functions. See :pull:`553`.
+- :func:`~hydra_zen.builds` and :func:`~hydra_zen.make_custom_builds_fn` now accept a `zen_exclude` field for excluding parameters from auto-population, either by name or by pattern. See :pull:`558`.
+
 
 .. _v0.11.0: